Reverted commit style
Both styles have their flaws. It really depends on if the database "going away" is more common than the script ending execution irregularly or timing out in a normal way. The latter is given some level of protection by the PDO methods.
Really hard to make a call on as I have not experienced either type of error so I cannot figure out their relative frequency or severity.